requirements.txt
Django==5.2.4
djangorestframework==3.16.0
python-dotenv==1.1.1
psycopg2==2.9.10
djangorestframework_simplejwt==5.5.0
django-filter==25.1